Simplify keystone OSA upgrade job config
The keystone OSA rolling-upgrade job defines a
builder and job template which duplicates things
done elsewhere. Instead of duplicating job
config, this patch implements the re-use of
other job config.

This also ensures that the primary checkout for
all openstack-ansible scenario jobs is actually
the openstack-ansible repository, instead of
another repository (eg: dragonflow). This ensures
that the correct tests are activated.

Change-Id: I8b0147454ba3ee1ecbd23a6f7c019826690936eb