Increase job timeout for openstack-ansible
With the demise of OSIC the openstack-ansible
integrated build is now being done more often
on much slower nodepool providers.

Also, recent work to implement rolling upgrades
have changed the way the initial deployment is
done and made it slower than before.

Unfortunately these two elements are now putting
us in the position where we're failing due to
a timeout more often than not, which is stalling
development.

This patch therefore increases the timeout by
30 mins.

Change-Id: I9075f06dc2e881dbc8642ada3f323edd522b9fae