Add support for integration testing ARA on different python versions
We've recently added py35 non-voting unit tests for ARA [1].
As we ramp up py35 support and testing, it's important to be able to
run integration tests on both py27 and py35 as well.

In order to limit the amount of jobs this would end up running because
we're interested in testing under both CentOS and Ubuntu, we'll compromise
by running py27 jobs on CentOS and py35 on Ubuntu.
We can rebalance the jobs running once Ansible 2.4 is out and 2.1 goes
EOL.

Since py35 is not yet passing for ARA and Ansible itself is working on
full py3 support, only start testing from Ansible 2.3+ and make them
non-voting at first.

[1]: https://review.openstack.org/#/c/481823/

Change-Id: I01e9e343f8f4416cdff5a653260fb2c168175a79