jenkins: openstack-ansible: Add openSUSE 42.3 jobs
openSUSE 42.3 is being released soon but we are already building images
for it so add it as a non-voting job to the OpenStack Ansible roles.
However, we don't add it to the AIO project yet as that still doesn't
work properly on 42.2 so it's pointless to add 42.3 at this point.

Change-Id: I9a5ec74eca996733ac97cca284f8e20f6886ee13