Add support for post-processing stestr runs to run_tox.sh
stestr [1][2][3] is a fork of the testrepository project that
concentrates on being a dedicated parallel python test runner. It fixes
a bunch of bugs and adds a bunch of features on top of testrepository.
Moving forward several projects will be migrating to using it instead of
testr by itself. (long term I'd like to get everything over to use
stestr) The plan is to also switch os-testr under the covers
eventually. To support the migration in the future this commit makes
sure the post processing steps we do in run_tox.sh can handle runs from
stestr as well as testrepository. The basic flow is the same but the
command and repository paths are slightly different.

[1] https://github.com/mtreinish/stestr
[2] http://stestr.readthedocs.io/en/latest/
[3] https://pypi.python.org/pypi/stestr

Change-Id: I1fc3b146e99e3484fa009b227ee5d0aeb6ed6084