Add pre-release and release automation for ARA
I managed the pre-release and release of ARA manually although I
pushed the signed tags to gerrit properly.
I have been taught there was a better way and that I should let
Zuul handle this for me.

That's an awesome idea. Let's do it.

Casualty of this change is the post coverage job which is no longer
necessary.

Change-Id: I5693994f1012a99f3b988db1c76b73d5ded711fd