Remove OSIC from the nodepool pool
Sadly the time has come where the OSIC cloud is needing to be fully
decom'd. It's with great sadness that write this commit message
and submit this PR however, as everyone is well aware, the OSIC
agreement between Rackspace and Intel has been discontinued for
some time and while Rackspace was able to keep the environment
online through feature freeze we must now sunset the environment
effective immediately.

I want to personally thank everyone within the community,
specifically the infrastructure team, for their time and effort
in using and supporting this source based publicly accessible
private cloud. The learnings we've had from this environment
have been invaluable and IMHO we've helped strengthen our
respective communities (deployer, operator, developer, users).
For me personally, It's been a wonderful experience building
and operating the OSIC and it's my hope that one day we can
do something like this again.

Change-Id: I2d1b0710e875bd1ebc305fb5b184b68bf18f2ef7
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>