Publish placement-api-ref
Make a special placement-api-ref publishing job for the
placement API. A special job is required because placement is hosted
within the nova repository, in its own sub-directory. The api-ref template
job is not set up for that.

We could probably have done some refactoring to reuse the existing
template but since it is planned for placement to eventually move out of
the nova repo, it's not really worth it.

This change also makes the gate-placement-api-ref job voting, the change
linked via depends-on below provides testing to ensure that all
placement URLs and methods have documentation.

In I895333c21a18afadaf987f8f9903a0c8b132b1d5 there was discussion that
this style of publishing job was incorrect and api-ref should be moved
under doc/source. This is not desirable for API reference because we do
not wish for API references to be versioned in the same way that project
documentation is versioned. Also, placement's odd positioning within
a repo with another service makes that complex. This change continues
the desirable publishing to developer.openstack.org and maintains the
patterns and standards set by nova's api-ref.

Co-Authored-By: Andrey Volkov <avolkov@mirantis.com>
Depends-On: Ifefedcd51f1f8d5e741ef6d02b7a0879a185e05f
Change-Id: If5ff9bcc31e2680861e959ef7dc8581210e619e9