jenkins: jobs: ansible-role-jobs: Increase job timeout to 90 minutes
We are seeing a lot of failures lately due to job timeouts. It appears
that jobs got more complex over time so these timeouts do not
necessarily mean that the job is stuck. For example the following job

2017-08-15 15:08:55.607415 | [Zuul] Launched by zl07
[...]
2017-08-15 16:03:01.405355 | TASK [Run functional tests for swift] ******************************************
2017-08-15 16:03:01.428393 | Tuesday 15 August 2017  16:03:01 +0000 (0:00:00.062)       0:51:42.612 ********
2017-08-15 16:14:02.430223 | [Zuul] Job timed out, result: FAILURE

took about an hour to reach the point of actually executing the
functional tests so the failure does not look legit. As such it might
be better to increase the timeout again so we can give the CI jobs more
time to finish.

Change-Id: I3562263beb5d3b8b3e164f96ae48f2cb2dddace5