Add Python 3.6 tox jobs
Add a sample python 3.6 job running on a Fedora 26 node.

Create the job macros and definitions for a Python 3.6 job running on
Fedora 26.  Add to the low-turnover bashate project so we can run some
tests on a real project.

At this point, I have not templated it out futher.  We could add it to
templates like python-jobs and create groups in zuul/layout.yaml to
make it eaiser if required.

Change-Id: I04fab592032621325f8638680fc5c842f316684a