Consume upload-pypi role
We created a new role in zuul-jobs, so lets start using it.

Make the job playbook such that it can just use named secrets so that we
can have a test job that uses the same playbook just with a different
secret. This should also allow someone to use the same job with their
own credentials if they wanted to.

Change-Id: I9440f54a40937bab6a688f14715e038080d6e944
Signed-off-by: Paul Belanger <pabelanger@redhat.com>