Remove/migrate Fedora 25 jobs
Fedora 26 is working with devstack jobs and (with bindep 2.5.0) tox
testing.  We wish to stop Fedora 25 builds to reduce the daily upload
requirements (but also, we try to keep this on the most recent as
quickly as possible).

The bindep fallback tests are removed and the other jobs migrated to
fedora-26 nodes.

Nodes are reduced to 0 in nodepool for build removal

Change-Id: I7bf569c5d7196516c93f8a2b7306596a74eaa0c7