Remove Packaging Deb project from gate
This is the first commit in a series to retire the Packaging
Deb project. Step 1 is to end project gating (as in
https://docs.openstack.org/infra/manual/drivers.html#retiring-a-project).

There are some lingering remnants in
jenkins/jobs/packaging-mirror.yml, a publisher 'deb-pkgs'
jenkins/jobs/macros.yml, and noop-jobs in zuul/layout.yml,
which I left for the Step 3 commit to entirely remove the
project from OpenStack Infra.

Change-Id: I53b200dd63b704cbf6f75225d8f8aca8d4fbcbe0