Add oaktree to v3 project list for readthedocs testing
Non-official projects publish docs with readthedocs, but none of our
current active v3 projects use it. oaktree is mostly currently inactive
but does use the hook-oaktree-rtfd job, so can be a useful test to
verify the new trigger-readthedocs job works.

Remove docs-on-rtfd from oaktree in v2 config so that we can verify that
it was v3 that triggered the readthedocs build.

Change-Id: Iee3f560c83773294b0dc350c582627ba4b5f99e5
Depends-On: I8c576c7a26ab123bd70f7a025d8a49a5cb3e2432