Switch puppet-integration-jobs for puppet-ceph to non-voting
The latest OpenStack release on Ubuntu ships with 12.X which
requires the ceph-mgr to run.  There is a patch pending to add
it however it cannot merged without a patch in both p-o-i and
puppet-ceph which is not possible due to CI blocking.

This patch will be later reverted once the puppet-ceph fix
lands.

Change-Id: Ib979a44e9fb20469332ed16cb284383198a086d6