Exclude upper constraint updates for driverfixes branches
Driverfixes branches are meant to track their parent stable branches
and are supposed to be limited in scope to just driver fixes past
normal stable policy.

Since requirements are not branched along with these branches, there
is no need to have a change proposed to switch their upper constraints
to point to a driverfixes/ version as they just get abandoned.

Change-Id: I53e9a1dc20c7c4af1e67b640835b27e05a0d5ebf