Add translation jobs
Add upstream_translation_update and propose_translation_update jobs.
They are close enough to put them both in the same commit.

Adds a legacy role to prep the upper-constraints file for tox.
Eventually this should work like the tox role, as an env variable passed
along for tox to read.

Adds a role to prep zanata client on the node in order to use the client
to interact with the Zanata server.

Change-Id: If736642860c5b0d1523215f90c64e858387d4254
Depends-On: I014d559c330803213105ce1bf83a41a3c157df9c
Signed-off-by: Jesse Keating <omgjlk@us.ibm.com>