Remove ironic nv job from tempest check pipeline
Below ironic job runs in Tempest check pipeline
-gate-tempest-dsvm-ironic-ipa-wholedisk-bios-agent_ipmitool-tinyipa-ubuntu-xenial-nv

This might be added when Tempest used to have Ironic tests but now
all ironic tests have been moved to Ironic tempest plugin.

As Tempest does not have any tests, it is better to remove this job
from Tempest pipeline.
This job runs on Nova, neutron as nv and in Ironic as voting.

Change-Id: If1a83713276dd54a8ca2f89a4772b16539fbc79c