Add openSUSE Tumbleweed support to the unbound element
In Tumbleweed or newer the /etc/rc.d (deprecated) directory
was dropped so we ended up failing in the wrong case of
the pile of workarounds. With this patch we use the recommended
location for any opensuse distribution, which should
be fine for all opensuse flavors currently in the gate.

Change-Id: Ibd690aff18f2acca310e3beb900ce2f471c2f772