Make a publish-docs-draft base job
We need a few different jobs that publish to docs-draft. Make a base
job that can be used for that purpose. This lets us move the definition
of openstack-doc-build back to openstack-zuul-jobs if we want, or we can
leave it where it is.

Use the same playbooks as for publish-openstack-docs. The only
difference was that the docs-draft stuff was fetching to /docs-draft/
instead of to /artifacts/ - and that's just a temporary location.

Move the tox envlist and command line into the run playbook. We don't
want those overridden. However, whether or not this uses constraints
(openstack does, infra doesn't) needs to be settable in child jobs.

Change-Id: I7737235ae265bcd19616adec3be921b78df579c4