Don't remove puppet in puppet prep macro.
We are running into errors when attempting to uninstal hiera on trusty
jobs because the hiera package does not exist. We no longer install
puppet as part of our image build process so we shouldn't need to
uninstall puppet, hiera, and facter at all. Simply drop the the removals
to address this as there is nothing to remove any longer.

Change-Id: I1ce79f3f7dd164e0635a4ae45210194f50ab8ad9