Add main.yaml tool
Add a tool to fill in missing bits in main.yaml from projects.yaml.

Use ruamel.yaml instead of pyyaml so that comments are preserved. It has
configurable indentation settings too - but the one with the format we
want adds 2 initial spaces to everything because the top element is a
list. (Well, except for the comment lines) So we still have to do a hack
at the end to make it suck less - but it's not MUCH of a hack, so
whatever.

Change-Id: I5e2da6f47fb7196c64c0c1b476d7e58c4d76be81