Fix publish infra docs index job
This job, like the specs site index, uses a custom tox env rather
than the normal way docs are built (largely on account of
project-config not having a setup.py).  It writes the output to
a custom location as well.  We are unable to use the normal infra
afs docs playbooks due to that, as well as the branch switching
logic not applying.  Let's drop the infra-docs post playbook
from this job and add the only parts we need from it, the
root-marker and fetch roles, to the post playbook unique to this
job.

Change-Id: I20db51fe80c633a217cbdc1f7bd7f8a47af27002
Depends-On: Ia6c6257cb1479abef79bf919cb6ad25ea6828d90