Fix bad python35 jobs migration
Something went sideways and standalone python35 jobs got transitioned
non-voting legacy-python35 jobs.

This fixes them to be what they should have been - with the following
logic. If non-voting legacy-python35 exists in both check and gate, then
it was originally voting (because we don't let non-voting things into the
gate normally) If it was only in check, it was non-voting.

We use the templates for this, which is what was originally supposed to
happen.

Change-Id: Iad3af8b526e23bd84db9574a939ad3ecbcef7296