Change the docs job to a deploy-publish-job
The charm-deployment-guide project isn't a 'typical' openstack project
in that it doesn't have a docs for the project, and a separate
deploy-guide.  The project IS a deploy-guide.  A separate change will be
needed in the charm-deployment-guide to move the docs to /deploy-guide.

This patch switches the job so that it actually does deploy-guide-jobs
instead.

Note:
 * This patchset supports the zuul v3 jobs

Change-Id: I74312e4dd22ffad4b0102b825c13e5f5e1c72ee8