Create openstack-sphinx named publish jobs
The build job is build-openstack-sphinx-docs but we never renamed the
publish jobs. Make new copies that are named correctly and update
project-config. We'll need to land a change to openstack-zuul-jobs
consuming the new name in project-templates before we can remove the old
job name.

While we're in there, rename the infra index job.

Change-Id: I14191ef736b35beb4e83005e760dff784ace1d5d