Add releasenotes publication job
We need to publish releasenotes to docs/releasenotes, so we need a new
publication playbook.

While doing this, it became clear that doc publication jobs were broken
(the infra ones were correct) The publish playbook looks for docs in
artigfacts, but the fetch was putting them into logs like we do for
builds.

Depends-On: I0ba254501e9f913fd0aff8a03f1b4f127560c759
Change-Id: I3e484efba79878ec460e124f611febf0d7454ce3