Use publish-to-pypi and friends for python releasing
We have new jobs and new project-templates for this but the migration
script borked them. Use them.

There are three templates, publish-to-pypi, publish-to-pypi-quietly and
publish-xstatic-to-pypi. publish-to-pypi-quietly is for projects that
were not doing the release announcement job.

Additionally, there were double-adds of
publish-openstack-python-tarball, including one with a dependency on
itself. Normalize those down to a single instance if it's in the job and
the job is only doing that and not one of the other publication
templates.

Change-Id: Ie8876fec55830e26ac39ccf91c9a3265604ae56b