Create static publication base job
This is for projects that need to publish something to
static.openstack.org. It publishes to
 /srv/static/{{ zuul.project.short_name }}
so that it's safe to use wherever.

There are four others that will need their own jobs since they need to
publish to somewhere other than zuul.project.short_name.

Also, fix yaml2ical publication since copy-pasta-ing from it discovered a bug.

Change-Id: Ie73bbb0c808c2e99f450dc0bbed4b0c4aa2b9fd4