Prevent non-check changes, add infra pipelines
Change the pipeline config to include invalid requirements for all
non-check pipelines so that we can keep project-pipeline configs but not
have any changes be enqueued in them. This should allow running v2 and
v3 side by side.

Add infra- versions of gate and post so that we can continue to use v3
to gate zuul, zuul-jobs, openstack-zuul-jobs and project-config.

Change-Id: I084b819ea0a4b67f36df61ccb3bfd6963cc3940d