Switch infra repos to not use constraints
The Infra projects are not part of the OpenStack release and as such do
not follow upper-constraints. Switch them to the versions of the jobs
that do not apply constraints or add openstack/requirements to
required-projects.

release-tools was left alone since that's a release team repo and should
be the release team's choice.

Depends-On: I03cad569cc539becccc603a07b8de3e30007f2bb
Change-Id: I914f660dbcd9680f872ed99f380bda049e1e0d2c