Publish manuals
The documentation projects publish from publish-docs directory to
top-level of developer.o.o and docs.o.o sites.

The jobs here are used for openstack-manuals, security-doc and api-site
repositories.

Add new playbooks and jobs.

Add new secret for afs developer site, it's the same secret as used for
docs but with a different path.

These jobs are not used directly, they will be used in-repo.

Change-Id: I888883595af2fccd85aed2721276a468d506e6f3