Add an origin remote for tag-releases for notes
tag-releases needs to fetch notes. To avoid touching the jenkins script
too much while v2 and v3 co-exist, just add the origin remote for all
projects in the pre-playbook.

Once v3 is rolled out fully and there is breathing room, this should be
reworked along with the contents of the script to meet all of the use
cases in a more native manner.

Change-Id: I8224ec098756874d5422fdb119d72ea1509b72a3