Stop translation of trove
The current translation sync to trove is unmerged for two month, see
https://review.openstack.org/#/c/493290/ . During that time most of the
jobs failed. also, no commit was done for trove since September 15th.

Let's stop translation for the project and remove the jobs from the
system. We will freeze the content in zanata - and if there's in the
future again the option to translate trove, this change can be reverted.

Change-Id: Ic9046b0e0475b101f52bb298be241c9a1dc81ad3