Remove all OSA functional tests
In order to re-implement the functional tests using
the common openstack-ansible-tests repo, the legacy
tests need to be removed as they're currently not
working and blocking our ability to merge in most
repositories.

This cuts out all the legacy testing for the role
repositories only, leaving the legacy tests for
the 'openetsack/openstack-ansible' repo alone for
now, along with any tests implemented in other
non-OSA repositories.

Change-Id: Ic672ac977ee96becb0d26b3310d8865862b35456