import the clone_repo.sh script from the releases repo
Under zuulv3 the behavior has changed significantly to now require
specifying which repositories need to be cloned for each job, and
zuul-cloner only copies the repositories that are already
available. As soon as zuulv3 is released, our jobs relying on the old
fallback behavior of checking out from a cache when present but using
the upstream server in other situations will fail.

The new script clone_repo.sh handles the cases we have in the
tag-releases job. It is a copy of
openstack/releases/tools/clone_repo.sh from
Ic559d27881a9a89fb0dcb5295a4d7ed7c578112f. The duplication is expected
to be temporary because we want to move the tag-releases job
definition into the releases repository after the milestone.

The other changes update the shared function "clone_repo" to call the
script rather than zuul-cloner. This approach was taken to make it
clear that clone_repo.sh is a copy of the original script. The step to
update the list of tags in the local repo is redundant, so that is
also removed.

Change-Id: I294d7ec37a6ba570065b965dcb461bc020668ce4
Signed-off-by: Doug Hellmann <doug@doughellmann.com>