Remove legacy jobs from python-openstackclient
This only removed the jobs from python-openstackclient, other projects
still need to move them locally.

legacy-osc-dsvm-functional has been moved to the python-openstackclient
repo, the others are non-voting and the move to repo is being worked
in https://review.openstack.org/#/c/512680/.  Due to the races we see
in the underlying services, running thes jobs is only wasting resources.

Change-Id: I662c2e770137895212b7ce22278548bbd08fd655