Remove py35 based template from nodepool
This template causes tox-py35 to be run on every branch of nodepool.
There is no way to disable it in-repo because this adds a variant that
matches every branch, so it will always run. We haven't been running
py35 unit tests on nodepool master branch, so this is causing us to
not be able to merge to master. A subsequent change to nodepool will
add a job variant that will cause tox-py35 to run on the feature/zuulv3
branch.

Change-Id: I59f44c55225dd0587f4b80ed388199284c93febe