Revert "Add linters job to all OSA role jobs"
Our linters job relies on the presence of the 'openstack-ansible-tests'
repository, so it is better for us to rely on a child of the upstream
linters job in which we can specify the additional repo. This saves
having to do more adjustments in project-config in the future too.

A series of patches has been proposed to all OSA repositories to
replace these jobs with our own linters job.

This reverts commit 62001adadb2efc27bd3ae7ea5a8c10fcc623e69b.

Change-Id: I2c47ed8fdda015909420ea56990f94decbe8f308