replace the common git review configuration for release jobs
The common configuration is incorrect for the release jobs because it
uses the proposal-bot account instead of the release account.

While we're at it, this patch updates the ssh key used by copying the
one from the old static signing node to ensure that we have the right
key associated with the right user.

Change-Id: I309d4cfab92908190bdb6ee89fab4d247931c18e
Signed-off-by: Doug Hellmann <doug@doughellmann.com>