add required-projects for the release-openstack-python jobs
Several projects that depend on neutron or horizon for their testing
use a special install script run via tox that tries to use zuul-cloner
to make the dependency available before installing it from
source. Until we change the PTI for building sdists and wheels from
Python projects so that it does not use tox, we need to ensure those
repos are present for zuul-cloner to copy. Both openstack/neutron and
openstack/horizon can be removed from the list of required-projects
for this job when the job is modified to not use tox.

Change-Id: If6deeb6ce80e56a1e942dc539ba13bf2da3a4e63
Signed-off-by: Doug Hellmann <doug@doughellmann.com>