Remove py27-based template for nodepool
We don't want the tox-py27 job running on the feature/zuulv3
branch (we ended support for that a while back). The only way
to not make it run is to remove the openstack-python-jobs-no-constraints
template from the project definition and run that template's jobs
directly from the in-repo project definition. That change can be
seen in the referenced Needed-By change below.

Needed-By: I22007434b38129379690f4e469a1981ed7dcb68c
Change-Id: I65b2583a50effc7d743f2bb9231b35ec7659e4fa