Add translation semaphore
With currently 122 translation jobs that are all started at the same
time in the periodic queue, the translation server gives some random 429
failures and also the setup of some nodes fails downloading the zanata-cli
package.

Limit the number of current running translation jobs to 10 using a
semaphore for translation jobs.

Change-Id: Ic71764f20029db2fc55ab7cbb00204c80e975479