Use bash for the release/tag playbook
The tag playbook used for the tag-releases job attempts to export a
variable containing a path relative to the user's homedir. This
works fine for tilde expansion in bash, but in vanilla sh ends up
passing a literal tilde into the consuming scripts at least some of
which do not perform their own os.path.expanduser() on the string.
Go ahead and explicitly use bash for the shell task in this playbook
to get the advance expansion it seems to assume.

Change-Id: Id10e37e5a2a32b7051c986f0541b1ed9f326e95c