Add a Python release job for API services
OpenStack API services traditionally haven't published releases on
PyPI, but still want to follow all the other basic steps of the
release process aside from that (including tarball signing, release
announcements and so on). Add a specific job for this so we can
create a new project-template in a subsequent change to apply to
those projects.

Change-Id: I59b7b0af47a39ed8af599a162d8f80d7c2e5b187