Fix openstack-infra publishing
Add native zuul v3 job to publish index file of
openstack-infra/publications.
Run the post job on all branches, so that a change to the master branch
regenerates the index file.

Add native zuul v3 job to publish a specific publication from
openstack-infra/publications.
Use the job for the project.

Needed-By: I190b34217c87d45e13bca18c7e121189323b698f
Change-Id: I6da66649726a1a96d06d7ca643adeec8ea8649eb