Logstash jobs treat gz and non gz files as identical
We compress log files early so that we minimize network traffic and disk
storage requirements. But logically foo.txt and foo.txt.gz are identical
for our purposes when hosting log files. You can ask apache for either
and should get the same content.

This creates problems when indexing logs because we had been treating
the non .gz suffix'ed files as the logical names but now we don't have
them in all cases in ES because we submit jobs for the .gz file
resulting in the name in ES having the .gz suffix. Additionally if we
have both foo.txt and foo.txt.gz on disk when we generate the log
indexing jobs we'll submit two jobs to index the same log file contents.

We can address both these problems if we early on treat the non .gz as
the "logical" name and only submit jobs for the no suffix names.

Change-Id: Id99d6bc58f94c197b567ec4080ab24715128d855