Drop TripleO jobs from Puppet modules
TripleO Relies on package promotions for jobs to run but can build the
puppet-* packages as necessary for depends-on changes. The shared
tripleo queue as well as the excessive amount of resources taken
by this job configuration no longer justifies the additional coverage
they provide.  The Puppet OpenStack team believes removing these jobs
to be in the best intrest of the community to reduce the impact on the
upstream infrastructure as we already have enough coverage between
the existing puppet-openstack-integration and beaker jobs.

Additionally as TripleO is moving to containers and hopes
to remove these existing baremetal jobs during the Queens cycle. While
the container versions of TripleO jobs still use puppet today, they do
so in a fashion that is not necessarily supported by the Puppet
OpenStack community.

Co-Authored-By: Alex Schultz <aschultz@redhat.com>
Change-Id: I247993de1149cced97170363d97f066b2bc1203b