Re-order Zuul graphs
There is an easy correlation between the Executor load and the
amount of Running builds. By re-ordering graphs, we're able to get
them to show one directly above the other, making it easier to
spot trends in the data.

When they're in different columns, it's harder to see the
correlation.

Change-Id: I7706c3293fbd702695fb3e4e917e33ca947beef9