Revert "tripleo: index /var/log/*.log.txt files"
This reverts commit 1f4d096c346368d15869265e5a08ad93289e20dc.

We are reverting this as it appears to significantly add to the total
number of files being indexed and we are very behind. Also there appear
to be some bugs with this like indexing all of these log files as
console logs when many of them are openstack service logs.

Additionally this seems to have caused crm114 to fill up its disks as it
creates a new dir with learning data on disk for each unique file path.
We will need to update crm114 to better collapse files into things that
are less unique.

Reverting this allows us to take a step back and stabilize then we can
try to move forward with this more carefully ensuring that crm114 is
ready and that we know what we are indexing and apply the correct tags
to it.

Change-Id: Ie75d13f4275d74c7035e6b8f89bab47d1b053d03