Add some check for zuul.d/projects.yaml
Add framework to run checks on zuul.d/projects.yaml.

For now it checks that each project has the system-required template
and that zuul.d/projects.yaml is sorted.

Remove the file tools/zuul-projects_sorted.sh which did a sort test as
well.

Update list of repos so that they are in the order expected (bash and
python string comparison is different).

Change-Id: I29e0b7f49ec890f62d6c521e443248ab33cb819a
Depends-On: I19dbdebc1c02f0ae6ab6671b4cf3fb3c4382a820