Reduce rh1 max-servers to 60
We've been having pretty regular timeout issues in the jobs running
on this cloud.  When running at the previous max of 70 concurrent
jobs there is some performance degradation, which previously was
okay but now is pushing us over the timeout.  Rather than continue
to waste those jobs, it should be more efficient to reduce the
concurrency of the jobs and hopefully have them pass at a higher
rate.

Change-Id: I3a0beab2a881e96af848856d15d3a730d196db8f
Related-Bug: 1722864