Remove some non-standard things from bindep-fallback
Although we've removed RDO repositories, we've recently realised that
several of these packages come from the centos-release-openstack-*
repositories.  This repo was brought in with
Ia9c45b762eea9f2989c9c63fb944b9e5e1f17ed2 as part of install_puppet.sh
migrations.  We don't want this installed; specifically it makes
things hard for tripelo, but generically it's better for jobs to
explicitly depend on such repos.

There's already been some workarounds in here with SuSE, which doesn't
have liberasurecode (060b868520a83c422050c3d3a15c89bf52e93968) and so
skips it.  Since this is just a port of what was really ad-hoc
packages that made up a devstack node, we should take opportunities to
decrease anything in here to a LCD, rather than add workarounds.

To this end, this change removes:

 liberasurecode-*
 mongodb-*
 python-zmq
 redis
 zookeeper

from the fallback list.  None of these are available in base centos7.
As a consequence the skipping of liberasurecode and zookeeper on suse
and gentoo respectively is removed too.

At this point, any issues from removing these packages should be
easily handled by local bindep installs.

Change-Id: Iab9495bd381b135e8a0f2e8f35e51a9218cf8f40