Don't re-propose requirements changes that already exist.
Previously Gerrit would reject changes that were basically duplicates.
So when we submitted a duplicate requirements change it would not get
accepted. The new version of Gerrit now accepts these duplicates which
causes an excessive amount of patches to be posted.

Change the behavior so that know it will only propose a requirements
change when there is an actual change compared to previous proposed
patches.

Change-Id: I89a0477de4317a5cbe76d6ad3a069e49d22df0e2