Switch tripleo-ci scenario001 to non-voting
We're seeing excessive amount of failures on this job due what appears
to be a performance regression. As this has been an on-going issue for
over a week, we need to temporarily switch scenario001 to non-voting
while we continue to troubleshoot.  This change should be reverted ASAP
when the cause is identified and fixed.

Change-Id: Iabc01eb763428765436211be9e600cc59fce400e
Related-Bug: #1731063