Don't re-propose translation changes that already exist.
Previously Gerrit would reject changes that were basically duplicates.
So when we submitted a duplicate translation change it would not get
accepted. The new version of Gerrit now accepts these duplicates which
causes an excessive amount of patches to be posted.

Change the behavior so that know it will only propose a translation
change when there is an actual change compared to previous proposed
patches.

Change-Id: I1d94852265ba5898f7c0fd312221df7c469a5a25