Don't attempt to run py27 on nova-specs
With the move to zuulv3, the py27 job is never run on
nova-specs changes because the openstack-tox-py27 job
definition in openstack-zuul-jobs ignores changes to
docs and rst files, which is basically all that is in
the nova-specs repo. So we've dropped the py27 target
in nova-specs and just run the same tests under pep8.

Change-Id: I8300f3c9270095f9fd0ab0dbc8cd8155252d08ff
Depends-On: Ic87e79db6a8881398d670148b609e0b81658dfd1