Switch to pep8 jobs for cinder-specs
The check and gate jobs for cinder-specs were configured to run
py27 unit tests. There is no code in this repo, and the one unit
test we had was a crude attempt to do some rst validation of the
specs.

Since this job is configured to skip on doc only changes, that
caused nothing to really happen. The cinder-spec test we had has
now been removed, and the pep8 job has been updated to run doc8
against the spec directory. This switches the configured check
and gate jobs to instead run pep8 so we are able to catch and
prevent formatting errors in new specs.

Depends-On: Ida20764edde3a07c89703d82b41958c96548b239
Change-Id: I007bae790d8c54447cb7af6d468339be2ad092ba