Switch to pep8 jobs for manila-specs
With the move to zuulv3, the openstack-tox-py* job
definitions ignore doc only changes, so we have dropped
triggering doc checks from the py27,py35 targets and
are instead running doc8 against the spec directory from
the pep8 target.

Switch the configure check and gate jobs to run pep8
so that we can catch and prevent formatting errors in
new specs.

Change-Id: Ie01d83ab3c022057281b080f6822d644a0cf57e9
Depends-On: I82c8dbec713fc5575fd03c0ae290bf190bbdf8ca