Use the new sphinx PTI for special publish jobs
Both releasenotes and docs build jobs have updated PTI job content.

As a step to rolling it out, update the direct and infra jobs to use the
new direct sphinx method. Also add -python3 versions of the jobs for the
projects that need python3.

The publish-openstack-sphinx-docs job still uses tox. The next patch
will update it as well.

Change-Id: I2d0222225f4ae6c45d09f9352c256062e2535d63
Depends-On: Idd7caf7d88b56d61872906027b4ce7d743572ded