Temporarily disable releasenotes for aodh
We need to merge a fix for aodh releasenotes to fix broken RST.
Since this is on stable/ocata branch and we need additional
fixes on master branch for it, we cannot merge this right now.

So, disable releasenotes build temporarily to merge the stable/ocata
change - and then revert this change again.

Change-Id: I7fea15d9869ad862347317452b473b13fe2dcc36
Needed-By: If8df8a6e8cad2d9247f07c9ed7a71f29b19f7d51
Needed-By: I994aa3cf173ec8a9412239544e3269aca17adc25