Use only base publish-openstack-sphinx-docs job
The publish-openstack-sphinx-docs job is marked as final: true

The publish-openstack-sphinx-docs-[neutron|horizon] templates are
adding to required-projects, which is therefore silently failing as
the job can not be modified.  I think there was a zuul bug previously
that allowed this, which is probably why we didn't notice.

I think the job wants to be final: as it is doing the publishing.  I
also think the project shouldn't be required for doc publishing; we're
trying to get away from having dependencies for docs.  Convert these
to use the base job (even if the jobs fail, at least zuul will try to
run then ... currently the only message for these POST jobs is put
into the zuul logs silently).

Needed-By: Id1c6d334e314382bdff05b27e4a7421cfab0c94e
Change-Id: I57ace5de5fd36efe2e513674734e97a0048743b6